feat: Add spring-boot-starter-hateoas dependency and update EmployeeController
This commit adds the spring-boot-starter-hateoas dependency to the project's pom.xml file. It also updates the EmployeeController class to import the necessary classes from the org.springframework.hateoas package. Additionally, it modifies the all() method to return a CollectionModel of EntityModel objects instead of a List of Employee objects. These changes enable HATEOAS support in the application, allowing for the creation of hypermedia-driven RESTful APIs.